Move ESM JS code to a dedicated folder outside the test folder.
