+++ packages/addons/__tests__/my-first-test.js
@@ -15,7 +15,7 @@ limitations under the License.
 */
 
 import { describe, expect, test } from '@jest/globals';
-import { sum } from './sum.js';
+import { sum } from '../lib/sum.js';
 
 describe('sum module', () => {
   test('adds 1 + 2 to equal 3', () => {
