+++ packages/addons/lib/README.txt
@@ -1,0 +1,2 @@
+This is a temp directory for storing ESM files to validate that Jest can run ESM tests.
+
