+++ packages/addons/lib/sum.js
